refactor(vscode): extract keybinding map builder and share PLATFORM constant
- Extract buildKeybindingMap() from AgentManagerProvider into format-keybinding.ts
  (pure data transformation, no vscode dependency)
- Import shared PLATFORM constant from constants.ts in both AgentManagerProvider
  and fork-session.ts instead of duplicating the literal
- Reduces AgentManagerProvider.ts from 1922 to 1893 lines (under 1900 cap)
