+++ packages/kilo-vscode/src/agent-manager/constants.ts
@@ -10,6 +10,9 @@ import * as path from "node:path"
  */
 export const MAX_MULTI_VERSIONS = 4
 
+/** Telemetry source identifier for all Agent Manager events. */
+export const PLATFORM = "agent-manager" as const
+
 /** Kilo config directory name (project-level and inside worktrees). */
 export const KILO_DIR = ".kilo"
