+++ packages/kilo-vscode/src/agent-manager/fork-session.ts
@@ -2,8 +2,7 @@ import type { KiloClient, Session } from "@kilocode/sdk/v2/client"
 import { getErrorMessage } from "../kilo-provider-utils"
 import { TelemetryProxy, TelemetryEventName } from "../services/telemetry"
 import type { WorktreeStateManager } from "./WorktreeStateManager"
-
-const PLATFORM = "agent-manager"
+import { PLATFORM } from "./constants"
 
 export interface ForkContext {
   getClient: () => KiloClient
